docs(core): static-query migration should use permalink for migration guide (#30649)
2d263f115065d550dd1874c3b5563f37b40ac9c2 updated the static-query migration
to refer to the new guide on AIO. Unfortunately these URLs are currently not
valid as the guide is only available on `next.angular.io` right now. In order to
make the link work permanently (e.g. if we eventually remove the guide in future
major versions), we use the permalink from the `v8` subdomain.

PR Close #30649