fix(ivy): correct timing of NgModuleFactory registration (#30706)
Commit 285a5a5f6 introduced registration of NgModules with ids when compiled
with AOT, and 8502ee6cf corrected the timing to avoid issues with tree
shaking. Neither of these approaches were correct.

This commit fixes the timing to match View Engine and avoid tree shaking
issues, as well as fixes a bug with the registration of imported module ids.

A new Ivy-only test is added which verifies that modules get registered
correctly under real-world conditions.

PR Close #30706