ci: re-enable chrome mobile tests on android (#32447)
With 1f33dcd28ff146c35f713b4cde2b3e70f6680d1b, we temporarily disabled
the chrome mobile tests on Android because the launched browsers weren't
able to load the karma page through the tunnel.

The Saucelabs support responded to a ticket that we opened and said that
there were issues with those browsers which are now resolved.

PR Close #32447