ci: disable hammerjs tests of components repository (#32485)
Follow-up for 42315c89e5d678ec8b1268a5cfae8d7c7069cfc9. We need to
disable a few HammerJS gesture tests of the component repository by
adding them to the material-ci test blocklist.

This is now surfaces because we updated the commit of the
`material-unit-tests` job to a more recent state where we re-enabled
the HammerJS gesture tests on the components-side. The tests were
previously disabled on the components repository because the blocklist
didn't work on Angular. See:
https://github.com/angular/components/commit/eaf70ca2a0600757041633976b29ab5a95d08296.

PR Close #32485