build: remove unused `polyfills-runtime.ts` file (#34424)
The `polyfills-runtime.ts` file is used in the [integration/ivy-i18n][1]
project, which has an appropriate [configuration][2]. The file was
accidentally included in the `cli-hello-world-ivy-i18n` integration
project was introduced in 0e38e26eb, although it is not used there.

This commit removes th `polyfills-runtime.ts` file from the
`cli-hello-world-ivy-i18n` integration project.

[1]: https://github.com/angular/angular/blob/8c5cb4b3f/integration/ivy-i18n/src/polyfills-runtime.ts
[2]: https://github.com/angular/angular/blob/8c5cb4b3f/integration/ivy-i18n/angular.json#L65-L72

PR Close #34424