Revert "refactor(compiler): refactor and export Bazel host creation. (#34686)" (#35063)
This reverts commit 7d626dd8aef6242dd3c8a8556fcec3b12b91df67.

The intention of this commit was for a consumer of the `compile` function to
pass the `bazelHost` it returns into future invocations, reusing the
`FileCache` between them. However, first-party ngc_wrapped does not do this,
which caused a performance regression as the `FileCache` was no longer
shared between compilations.

PR Close #35063