fix(ngcc): do not warn if `paths` mapping does not exist (#36525)
In fa7ab6d0be816a82f71aae96a6c32dc9da0eb876 the `getBasePaths()`
function was changed to log a warning if a `basePath()` computed from
the `paths` mappings did not exist. It turns out this is a common and
accepted scenario, so we should not log warnings in this case.

Fixes #36518

PR Close #36525