ci: update tested browsers to match supported browsers (#36973)
With a previous change we add FireFox Extended Support Release (ESR) to the list of supported browsers.

With this change we now add this above mentioned browser to list of tested browsers and also update several other browsers.

Removed browsers because they are no longer supported:
- iOS 10 and 11
- Safari 10 and 11

Added
- Firefox 68 (Current ESR version)
- Safari 12 and 13
- iOS Safari 12 and 13

Updated
- Update Chrome to 81
- Updated Firefox to 76

PR Close #36973