refactor(docs-infra): remove `rev` property from `Resource` interface and data (#37181)
The `rev` property has been in the initial commit that introduced
`resources.json` (edc1cbdff7ea805121ef7e77b29c08c2b0074290) and has been
added to all new entries since (always with the value `true`). This
field is a remnant from back when this data was stored in a Firebase
Database and the `rev` field indicated whether the entry has been
reviewed/approved by a DevRel lead or something. Now that the data is
kept in the repository and the reviewing is done as part of the
corresponding PR, this field is no longer necessary.

PR Close #37181