refactor(compiler-cli): remove event output helper from TCB (#40738)
In 84220dc7a689cffb75df4b711b9477ef38bdd12b the `EventEmitter.subscribe`
API was extended with a new signature that allows the emitter's generic
type `T` to flow into the subscribe callback. This new signature removes
the need for the special `_outputHelper` function that used to be
emitted into TCBs when `strictOutputEventTypes`/`strictTemplates` is
enabled.

PR Close #40738